requirements.txt
requests
beautifulsoup4
selenium
pandas
numpy
openai
chromadb
python-dotenv
streamlit     